requirements.txt
flask>=3.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
mediapipe>=0.10.0
ultralytics>=8.1.0
gunicorn>=21.2.0
python-dotenv>=1.0.1
werkzeug>=3.0.1
psycopg2-binary>=2.9.9
cloudinary>=1.39.0
